red-v1.1-trials/portal/requirements.txt
libvirt-python
kagglehub